addressed Vanaja comments
